third_party/WebKit/LayoutTests/external/wpt/web-animations/timing-model/animation-effects/local-time-expected.txt

This is a testharness.js-based test.
FAIL Local time is current time for animation effects associated with an animationassert_equals: expected (number) -1000 but got (object) null
PASS Local time is unresolved for animation effects not associated with an animation
Harness: the test ran to completion.
